Being refactoring stack_analyzer for libClusterFuzz.
- Move majority of code to lib/clusterfuzz/stacktraces.
- Rename StackAnalyzerState to stacktraces.CrashInfo.  Also move several
  attributes that are better suited to the StackParser class.
- Introduce StackParser class to better store/keep state.
- stack_analyzer becomes an embedder, and does some additional steps
  such as symbolizastion and Android linkification.
- Remove StackAnalyzerState.tool and tool_from_group functionality. They
  were unused.
