+++ src/python/lib/README.md
@@ -1,0 +1,1 @@
+# libClusterFuzz
